fixed: add switch network component
Closes #973
